adsada

+++ app/Http/Controllers/AreaOneController.php
@@ -162,8 +162,10 @@ class AreaOneController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->email == 'chhattofficial@chhatt.com'){
         $item = AreaOne::find($id);
         $item->delete();
+        }
         return redirect()->back();
     }
 }
