+++ app/Http/Controllers/CityController.php
@@ -109,8 +109,10 @@ class CityController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->email == 'chhattofficial@chhatt.com'){
         $item = City::find($id);
         $item->delete();
+        }
         return redirect()->back();
     }
 }
